feat(TTC-57): refactored HomePage to new structure
