+++ src/components/templates/HomeLayout/HomeLayout.tsx
@@ -1,26 +1,16 @@
 import type React from 'react';
-import type { CategoryBanner } from '../../../types/Category';
-import { ShopByCategory } from '../../organisms/ShopByCategory';
 import styles from './HomeLayout.module.scss';
 
 type Props = {
-  shopByCategoryBanners: CategoryBanner[];
-  isLoadingCateogriesCount: boolean;
+  shopByCategorySection: React.ReactNode;
 };
 
-const HomeLayout: React.FC<Props> = ({
-  shopByCategoryBanners,
-  isLoadingCateogriesCount,
-}) => {
+const HomeLayout: React.FC<Props> = ({ shopByCategorySection }) => {
   return (
     <>
       <h1 className={styles.homeTitleExample}>Home Page</h1>
-
       <h2>Shop by category</h2>
-      <ShopByCategory
-        isLoading={isLoadingCateogriesCount}
-        categories={shopByCategoryBanners}
-      />
+      {shopByCategorySection}
     </>
   );
 };
