+++ src/pages/HomePage/HomePage.tsx
@@ -1,24 +1,8 @@
+import { ShopByCategory } from '../../components/organisms/ShopByCategory';
 import { HomeLayout } from '../../components/templates';
-import { staticCategoryData } from '../../data';
-import type { CategoryBanner } from '../../types/Category';
-import { useProductCountsByCategory } from '../../hooks';
 
 const HomePage = () => {
-  const { data: counts, isLoading } = useProductCountsByCategory();
-
-  const shopByCategoryBanners: CategoryBanner[] = staticCategoryData.map(
-    (category) => ({
-      ...category,
-      productCount: counts?.[category.categorySlug] || 0,
-    }),
-  );
-
-  return (
-    <HomeLayout
-      isLoadingCateogriesCount={isLoading}
-      shopByCategoryBanners={shopByCategoryBanners}
-    />
-  );
+  return <HomeLayout shopByCategorySection={<ShopByCategory />} />;
 };
 
 export default HomePage;
